requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
cryptography
plaid-python>=18,<37
